Refactor package structure and add unit tests for Teradata integration
