+++ plugin/trino-teradata/src/test/java/io/trino/plugin/integration/TeradataJdbcConnectorTest.java
@@ -73,6 +73,15 @@ public class TeradataJdbcConnectorTest
         requireNonNull(verifyResults, "verifyResults is null");
         requireNonNull(verifyFailure, "verifyFailure is null");
         QueryAssertions.QueryAssert queryAssert = Assertions.assertThat(queryAssertProvider);
+
+        try {
+            queryAssert.result();
+        }
+        catch (Throwable var5) {
+            verifyFailure.accept(queryAssert.failure());
+            return;
+        }
+
         verifyResults.accept(queryAssert);
     }
 
@@ -148,13 +157,11 @@ public class TeradataJdbcConnectorTest
         return OptionalInt.of(TERADATA_OBJECT_NAME_LIMIT);
     }
 
-    @Override
     protected void verifySchemaNameLengthFailurePermissible(Throwable e)
     {
         assertThat(e).hasMessage(format("Schema name must be shorter than or equal to '%s' characters but got '%s'", TERADATA_OBJECT_NAME_LIMIT, TERADATA_OBJECT_NAME_LIMIT + 1));
     }
 
-    @Override
     protected OptionalInt maxColumnNameLength()
     {
         return OptionalInt.of(TERADATA_OBJECT_NAME_LIMIT);
@@ -166,27 +173,16 @@ public class TeradataJdbcConnectorTest
         assertThat(e).hasMessageMatching(format("Column name must be shorter than or equal to '%s' characters but got '%s': '.*'", TERADATA_OBJECT_NAME_LIMIT, TERADATA_OBJECT_NAME_LIMIT + 1));
     }
 
-    @Override
-    @Test
-    public void testDataMappingSmokeTest()
-    {
-        // Skipping the Data Mapping smoke test as this is consuming more time to complete all data type mapping on Teradata ClearScape instance. Will enable once we fix timeout error when running tests on Teradata ClearScape instance
-        skipTestUnless(false);
-    }
-
-    @Override
     protected OptionalInt maxTableNameLength()
     {
         return OptionalInt.of(TERADATA_OBJECT_NAME_LIMIT);
     }
 
-    @Override
     protected void verifyTableNameLengthFailurePermissible(Throwable e)
     {
         assertThat(e).hasMessageMatching(format("Table name must be shorter than or equal to '%s' characters but got '%s'", TERADATA_OBJECT_NAME_LIMIT, TERADATA_OBJECT_NAME_LIMIT + 1));
     }
 
-    @Override
     @Test
     public void testDistinctLimit()
     {
@@ -200,7 +196,6 @@ public class TeradataJdbcConnectorTest
 
     /* Overriding the method as Teradata avg calculations are slightly different than trino so Skipping the results check for avg
         Expecting actual: (111.660, 111728394.9938271616, 1.117283945E8, 111.6605) to contain exactly in any order: [(111.661, 111728394.9938271605, 1.117283945E8, 111.6605)] */
-    @Override
     @Test
     public void testNumericAggregationPushdown()
     {
@@ -234,7 +229,6 @@ public class TeradataJdbcConnectorTest
     }
 
     // Overriding this test case as Teradata defines varchar with a length.
-    @Override
     @Test
     public void testVarcharCastToDateInPredicate()
     {
@@ -301,7 +295,6 @@ public class TeradataJdbcConnectorTest
     }
 
     // Overriding this test case as Teradata raises different error message for division by zero.
-    @Override
     @Test
     public void testArithmeticPredicatePushdown()
     {
@@ -317,7 +310,6 @@ public class TeradataJdbcConnectorTest
         }
     }
 
-    @Override
     @Test
     public void testCreateTableAsSelect()
     {
@@ -382,7 +374,6 @@ public class TeradataJdbcConnectorTest
     }
 
     // Overriding this test case as Teradata does not support negative dates.
-    @Override
     @Test
     public void testDateYearOfEraPredicate()
     {
@@ -390,7 +381,6 @@ public class TeradataJdbcConnectorTest
     }
 
     // Override this test case as Teradata has different syntax for creating tables with AS SELECT statement.
-    @Override
     @Test
     public void verifySupportsRowLevelUpdateDeclaration()
     {
@@ -405,7 +395,6 @@ public class TeradataJdbcConnectorTest
 
     // Override this test case as Teradata has different syntax for creating tables with AS SELECT statement.
     // TODO Will handle this while Teradata connector supporting WRITE operations.
-    @Override
     @Test
     public void testJoinPushdown()
     {
@@ -453,7 +442,6 @@ public class TeradataJdbcConnectorTest
         }
     }
 
-    @Override
     @Test
     public void testCharVarcharComparison()
     {
@@ -567,7 +555,6 @@ public class TeradataJdbcConnectorTest
     }
 
     // Overriding this test case as Teradata doesn't have support to (k, v) AS VALUES in insert statement
-    @Override
     @Test
     public void testVarcharCharComparison()
     {
@@ -581,11 +568,11 @@ public class TeradataJdbcConnectorTest
     }
 
     // Overriding this test case as Teradata supports timezone in different way.
-    @Override
     @Test
     public void testTimestampWithTimeZoneCastToDatePredicate()
     {
         skipTestUnless(this.hasBehavior(TestingConnectorBehavior.SUPPORTS_CREATE_TABLE_WITH_DATA));
+
         TestTable table;
         try {
             table = this.newTrinoTable("timestamptz_to_date", "(i varchar(20), t TIMESTAMP)",
@@ -602,7 +589,7 @@ public class TeradataJdbcConnectorTest
         TestTable e = table;
 
         try {
-            Assertions.assertThat(this.query("SELECT i FROM " + table.getName() + " WHERE CAST(t AS date) = DATE '2005-09-10'")).hasCorrectResultsRegardlessOfPushdown();
+            Assertions.assertThat(this.query("SELECT i FROM " + table.getName() + " WHERE CAST(t AS date) = DATE '2005-09-10'")).hasCorrectResultsRegardlessOfPushdown().skippingTypesCheck().containsAll("VALUES 'UTC', 'Los Angeles'");
         }
         catch (Throwable var7) {
             if (table != null) {
@@ -619,11 +606,11 @@ public class TeradataJdbcConnectorTest
         table.close();
     }
 
-    @Override
     @Test
     public void testTimestampWithTimeZoneCastToTimestampPredicate()
     {
         skipTestUnless(this.hasBehavior(TestingConnectorBehavior.SUPPORTS_CREATE_TABLE_WITH_DATA));
+
         TestTable table;
         try {
             table = this.newTrinoTable(
@@ -642,7 +629,7 @@ public class TeradataJdbcConnectorTest
         TestTable e = table;
 
         try {
-            Assertions.assertThat(this.query("SELECT i FROM " + table.getName() + " WHERE CAST(t AS timestamp(0)) = TIMESTAMP '2005-09-10 13:00:00'")).hasCorrectResultsRegardlessOfPushdown();
+            Assertions.assertThat(this.query("SELECT i FROM " + table.getName() + " WHERE CAST(t AS timestamp(0)) = TIMESTAMP '2005-09-10 13:00:00'")).hasCorrectResultsRegardlessOfPushdown().skippingTypesCheck().containsAll("VALUES 'UTC'");
         }
         catch (Throwable var7) {
             if (table != null) {
@@ -659,11 +646,11 @@ public class TeradataJdbcConnectorTest
         table.close();
     }
 
-    @Override
     @Test
     public void testJoinPushdownWithLongIdentifiers()
     {
         skipTestUnless(this.hasBehavior(TestingConnectorBehavior.SUPPORTS_CREATE_TABLE) && this.hasBehavior(TestingConnectorBehavior.SUPPORTS_JOIN_PUSHDOWN));
+        String baseColumnName = "col";
         int maxLength = this.maxColumnNameLength().orElse(65541);
         String validColumnName = "z".repeat(maxLength - 5);
 
@@ -686,112 +673,90 @@ public class TeradataJdbcConnectorTest
         };
     }
 
-    @Override
     @Test
     public void testRenameSchema()
     {
         Assumptions.abort("Skipping as connector does not support RENAME SCHEMA");
     }
 
-    @Override
     @Test
     public void testColumnName()
     {
         Assumptions.abort("Skipping as connector does not support column level write operations");
     }
 
-    @Override
     @Test
     public void testCreateTableAsSelectWithUnicode()
     {
         Assumptions.abort("Skipping as connector does not support creating table with UNICODE characters");
     }
 
-    @Override
     @Test
     public void testUpdateNotNullColumn()
     {
         Assumptions.abort("Skipping as connector does not support insert operations");
     }
 
-    @Override
     @Test
     public void testWriteBatchSizeSessionProperty()
     {
         Assumptions.abort("Skipping as connector does not support insert operations");
     }
 
-    @Override
     @Test
     public void testInsertWithoutTemporaryTable()
     {
         Assumptions.abort("Skipping as connector does not support insert operations");
     }
 
-    @Override
     @Test
     public void testWriteTaskParallelismSessionProperty()
     {
         Assumptions.abort("Skipping as connector does not support insert operations");
     }
 
-    @Override
     @Test
     public void testInsertIntoNotNullColumn()
     {
         Assumptions.abort("Skipping as connector does not support insert operations");
     }
 
-    @Override
     @Test
     public void testDropSchemaCascade()
     {
         Assumptions.abort("Skipping as connector does not support dropping schemas with CASCADE option");
     }
 
-    @Override
     @Test
     public void testAddColumn()
     {
         Assumptions.abort("Skipping as connector does not support column level write operations");
     }
 
-    @Override
-    @Test
-    public void testDropNonEmptySchemaWithTable()
-    {
-        Assumptions.abort("Skipping as connector does not support drop schemas");
-    }
-
-    @Override
     @Test
     public void verifySupportsUpdateDeclaration()
     {
         Assumptions.abort("Skipping as connector does not support update operations");
     }
 
-    @Override
     @Test
     public void testDropNotNullConstraint()
     {
         Assumptions.abort("Skipping as connector does not support dropping a not null constraint");
     }
 
-    @Override
     @Test
     public void testExecuteProcedureWithInvalidQuery()
     {
         Assumptions.abort("Skipping as connector does not support execute procedure");
     }
 
-    @Override
     @Test
     public void testCreateTableAsSelectNegativeDate()
     {
         Assumptions.abort("Skipping as connector does not support creating table with negative date");
     }
 
-    @Override
     protected void assertCreateTableAsSelect(Session session, String query, String expectedQuery, String rowCountQuery)
     {
         String table = "test_ctas_" + TestingNames.randomNameSuffix();
@@ -810,7 +775,6 @@ public class TeradataJdbcConnectorTest
                 .build();
     }
 
-    @Override
     protected TestTable newTrinoTable(String namePrefix, @Language("SQL") String tableDefinition, List<String> rowsToInsert)
     {
         String tableName = "";
