+++ plugin/trino-teradata/src/test/java/io/trino/plugin/integration/clearscape/ClearScapeSetup.java
@@ -59,13 +59,8 @@ public class ClearScapeSetup
 
     public void cleanup()
     {
-        if (manager == null) {
-            return;
-        }
-        if (destoryEnv) {
+        if (destoryEnv && manager != null) {
             manager.teardown();
-            return;
         }
-        manager.stop();
     }
 }
