+++ plugin/trino-teradata/src/test/java/io/trino/plugin/integration/clearscape/Error4xxException.java
@@ -17,6 +17,11 @@ package io.trino.plugin.integration.clearscape;
 public class Error4xxException
         extends BaseException
 {
+    public Error4xxException(int statusCode, String body, String reason)
+    {
+        super(statusCode, body, reason);
+    }
+
     public Error4xxException(int statusCode, String body)
     {
         super(statusCode, body);
