--- plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/clearscape/BaseException.java
@@ -1,56 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.trino.plugin.teradata.clearscape;
-
-public class BaseException
-        extends RuntimeException
-{
-    private final int statusCode;
-
-    private final String body;
-
-    private final String reason;
-
-    public BaseException(int statusCode, String body)
-    {
-        super(body);
-        this.statusCode = statusCode;
-        this.body = body;
-        this.reason = null;
-    }
-
-    public BaseException(int statusCode, String body, String reason)
-    {
-        super(body);
-        this.statusCode = statusCode;
-        this.body = body;
-        this.reason = reason;
-    }
-
-    public int getStatusCode()
-    {
-        return statusCode;
-    }
-
-    public String getBody()
-    {
-        return body;
-    }
-
-    public String getReason()
-    {
-        return reason;
-    }
-}
