--- plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/clearscape/Constants.java
@@ -1,9 +1,0 @@
-package io.trino.plugin.teradata.clearscape;
-
-public interface Constants
-{
-    String ENV_CLEARSCAPE_URL = "https://api.clearscape.teradata.com";
-    String ENV_CLEARSCAPE_USERNAME = "demo_user";
-    String ENV_CLEARSCAPE_NAME = "trino-teradata-test-env";
-    String ENV_CLEARSCAPE_REGION = "asia-south";
-}
