--- plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/clearscape/Model.java
@@ -1,107 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.trino.plugin.teradata.clearscape;
-
-public class Model
-{
-    String envName;
-    String hostName;
-    String userName;
-    String password;
-    String databaseName;
-    String token;
-    String jdbcUrl;
-    String region;
-
-    public String getEnvName()
-    {
-        return envName;
-    }
-
-    public void setEnvName(String envName)
-    {
-        this.envName = envName;
-    }
-
-    public String getHostName()
-    {
-        return hostName;
-    }
-
-    public void setHostName(String hostName)
-    {
-        this.hostName = hostName;
-    }
-
-    public String getUserName()
-    {
-        return userName;
-    }
-
-    public void setUserName(String userName)
-    {
-        this.userName = userName;
-    }
-
-    public String getPassword()
-    {
-        return password;
-    }
-
-    public void setPassword(String password)
-    {
-        this.password = password;
-    }
-
-    public String getDatabaseName()
-    {
-        return databaseName;
-    }
-
-    public void setDatabaseName(String databaseName)
-    {
-        this.databaseName = databaseName;
-    }
-
-    public String getToken()
-    {
-        return token;
-    }
-
-    public void setToken(String token)
-    {
-        this.token = token;
-    }
-
-    public String getJdbcUrl()
-    {
-        return jdbcUrl;
-    }
-
-    public void setJdbcUrl(String jdbcUrl)
-    {
-        this.jdbcUrl = jdbcUrl;
-    }
-
-    public String getRegion()
-    {
-        return region;
-    }
-
-    public void setRegion(String region)
-    {
-        this.region = region;
-    }
-}
