--- plugin/trino-teradata/src/test/java/io/trino/plugin/teradata/util/TeradataTestConstants.java
@@ -1,22 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.trino.plugin.teradata.util;
-
-public interface TeradataTestConstants
-{
-    String ENV_CLEARSCAPE_URL = "https://api.clearscape.teradata.com";
-    String ENV_CLEARSCAPE_USERNAME = "demo_user";
-    String ENV_CLEARSCAPE_REGION = "asia-south";
-}
